requirements.txt
googlesearch-python
colorama
argparse
os